chore: move trace builders to the correct location
