+++ tests/e2e/utils/data_builder/__init__.py
@@ -5,7 +5,9 @@ from tests.e2e.utils.data_builder.metrics import (
     build_metric_name,
     build_metric_query_data,
     build_multiple_add_metric_input,
+)
+from tests.e2e.utils.data_builder.traces import (
     build_put_annotations_input,
     build_put_metadata_input,
+    build_trace_default_query,
 )
-from tests.e2e.utils.data_builder.traces import build_trace_default_query
